+++ gd-core/src/utils/array.js
@@ -1,11 +1,11 @@
 // @flow
 
-export const entries = <T>(obj: { [string]: T }): Array<[string, T]> => {
+export const entries = (obj: { [string]: any }): Array<[string, any]> => {
     const keys: string[] = Object.keys(obj);
     return keys.map(key => [key, obj[key]]);
 };
 
-export const values = <T>(obj: { [string]: T }): Array<T> => {
+export const values = (obj: { [string]: any }): Array<any> => {
     const keys: string[] = Object.keys(obj);
     return keys.map(key => obj[key]);
 };