+++ gd-gltg-bmp/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Added support for fetching config from the api
   [GLGVO-709](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-709)
+- Integrate the results view with `gltg-bmp` api
+  [GLGVO-711](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-711)
+
 
 ## [0.1.0] - 2020-10-01
 ### Added